requirements.txt
# Scraping
selenium>=4.0.0
webdriver-manager>=4.0.0

# PDF processing (Phase 3)
pymupdf>=1.23.0  # Faster and more powerful than pypdf

# OCR (Phase 3)
pytesseract>=0.3.10
pdf2image>=1.16.0
Pillow>=10.0.0
pandas>=2.0.0  # Required for OCR confidence analysis

# Database (using stdlib sqlite3 - no additional deps needed)

# Vector DB (Phase 5)
# qdrant-client>=1.7.0

# Embeddings (Phase 5)
# sentence-transformers>=2.2.0

# App (Phase 8)
# streamlit>=1.30.0

# Development
python-dotenv>=1.0.0
